+++ packages/utils/src/SchemaError/SchemaError.ts
@@ -1,4 +1,4 @@
-import type { v1 } from "@standard-schema/spec";
+import type { StandardSchemaV1 } from "@standard-schema/spec";
 
 /**
  * A schema error with useful information.
@@ -7,14 +7,14 @@ export class SchemaError extends Error {
   /**
    * The schema issues.
    */
-  public readonly issues: ReadonlyArray<v1.StandardIssue>;
+  public readonly issues: ReadonlyArray<StandardSchemaV1.Issue>;
 
   /**
    * Creates a schema error with useful information.
    *
    * @param issues The schema issues.
    */
-  constructor(issues: ReadonlyArray<v1.StandardIssue>) {
+  constructor(issues: ReadonlyArray<StandardSchemaV1.Issue>) {
     super(issues[0].message);
     this.name = "SchemaError";
     this.issues = issues;